+++ internal/acceptance_tests/blocks/dynamic_snippet_explicit_with_content.tf
@@ -1,0 +1,8 @@
+resource "fastly_service_dynamic_vcl_snippet" "test" {
+  service_id = fastly_service_cdn.test.id
+  version    = 1
+  name       = "{{.DYNAMIC_SNIPPET_NAME}}"
+  type       = "{{.DYNAMIC_SNIPPET_TYPE}}"
+  priority   = {{.DYNAMIC_SNIPPET_PRIORITY}}
+  content    = {{.DYNAMIC_SNIPPET_INLINE_CONTENT}}
+}